+++ app/lib/schema-data.server.ts
@@ -15,7 +15,7 @@ import { getEnvOrDieInProduction } from './env.server'
 import type {
   ReferencedSchema,
   Schema,
-} from '~/routes/docs_.schema.$version.$/components'
+} from '~/routes/docs_._schema-browser.schema.($version).$/components'
 
 const GITHUB_API_TOKEN = getEnvOrDieInProduction('GITHUB_API_TOKEN')
 const octokit = new Octokit({ auth: GITHUB_API_TOKEN })
