+++ app/routes/docs_._schema-browser.schema.($version).$/SchemaDocumentation.tsx
@@ -7,7 +7,7 @@
  */
 import { Link } from '@remix-run/react'
 
-export default function () {
+export function SchemaDocumentation() {
   return (
     <>
       <h2>Schema Browser</h2>
